#1 [SETTING] : add data plugin, replace feature plugin -> data at data and network module
