+++ build-logic/src/main/java/com/teamwable/wable/plugin/AndroidDataPlugin.kt
@@ -1,0 +1,23 @@
+package com.teamwable.wable.plugin
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.api.artifacts.VersionCatalogsExtension
+import org.gradle.kotlin.dsl.apply
+import org.gradle.kotlin.dsl.dependencies
+import org.gradle.kotlin.dsl.getByType
+
+class AndroidDataPlugin : Plugin<Project> {
+    override fun apply(target: Project) = with(target) {
+        plugins.apply("com.android.library")
+        apply<AndroidKotlinPlugin>()
+        apply<KotlinSerializationPlugin>()
+        apply<AndroidHiltPlugin>()
+        apply<RetrofitPlugin>()
+
+        val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
+        dependencies {
+            "implementation"(libs.findLibrary("timber").get())
+        }
+    }
+}
