+++ core/data/build.gradle.kts
@@ -1,6 +1,6 @@
 @Suppress("DSL_SCOPE_VIOLATION")
 plugins {
-    id("com.teamwable.wable.feature")
+    id("com.teamwable.wable.data")
 }
 android {
     namespace = "com.teamwable.data"
